Fix: Update instructions
Add pictures of different states of the game to tell the user what it
will look like. Rewrite instructions to match with the button names of
the game. Style headings to be bold so user can quickly find what they
are looking for.

Instructions before were vague and unclear. Updated them to be more
guideful for the user.